Refactor main.rb and app.rb to make the project more modular
