+++ Gemfile
@@ -1,5 +1,5 @@
 source 'https://rubygems.org'
 
 gem 'pry', '~> 0.13.1'
-gem 'rubocop', '>= 1.0', '< 2.0'
 gem 'rspec'
+gem 'rubocop', '>= 1.0', '< 2.0'
