--- lib/app.rb
@@ -1,9 +1,0 @@
-class App
-  attr_accessor :books, :music_albums, :games
-
-  def initialize
-    @books = []
-    @music_albums = []
-    @games = []
-  end
-end
