+++ modules/books_services.rb
@@ -1,0 +1,6 @@
+module BooksServices
+  def add_book
+    puts 'Awesome Let\'s add a book'
+    sleep 0.5
+  end
+end
