+++ modules/games_services.rb
@@ -1,0 +1,6 @@
+module GamesServices
+  def add_game
+    puts 'Awesome Let\'s add a game'
+    sleep 0.5
+  end
+end
