+++ modules/music_album_services.rb
@@ -1,0 +1,7 @@
+module MusicAlbumServices
+  def add_music_album
+    puts 'Awesome Let\'s add a music album 🎶'
+    sleep 0.5
+    print 'Enter the name of the artist'
+  end
+end
